Refactor: Event와 Coupon 로직 분리
